cleaned up code, and renamed files to follow convention
