+++ src/reactviews/pages/ContainerDeployment/containerDeploymentStateProvider.tsx
@@ -3,11 +3,11 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as cd from "../../../sharedInterfaces/containerDeploymentInterfaces";
+import * as cd from "../../../sharedInterfaces/containerDeployment";
 
 import { useVscodeWebview } from "../../common/vscodeWebviewProvider";
 import { createContext } from "react";
-import { ContainerDeploymentContextProps } from "../../../sharedInterfaces/containerDeploymentInterfaces";
+import { ContainerDeploymentContextProps } from "../../../sharedInterfaces/containerDeployment";
 import { getCoreRPCs } from "../../common/utils";
 import { ConnectionGroupSpec } from "../../../sharedInterfaces/connectionGroup";
