+++ src/reactviews/pages/ContainerDeployment/prereqCheckPage.tsx
@@ -9,7 +9,7 @@ import { StepCard } from "./stepCard";
 import { Button } from "@fluentui/react-components";
 import { ContainerInputForm } from "./containerInputForm";
 import { checkStepErrored, isLastStepLoaded, runDockerStep } from "./deploymentUtils";
-import { DockerStepOrder } from "../../../sharedInterfaces/containerDeploymentInterfaces";
+import { DockerStepOrder } from "../../../sharedInterfaces/containerDeployment";
 import { ContainerDeploymentHeader } from "./containerDeploymentHeader";
 import { locConstants } from "../../common/locConstants";
 import { stepPageStyles } from "./sharedStyles";
